+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 Este arquivo refere-se às mudanças realizadas desde a versão inicial de desenvolvimento.
 
 * 1.0.6 - 10/10/2014
+  * Added - nova posição para módulos: bellow
   * Added - reescrita tpl_seb_blog sem contemplar todos os parâmetros
 * 1.0.5 - 02/10/2014
   * Added - reecrita de exibição para app work